Move maintenance mode to middleware
